+++ tests/test.js
@@ -1,7 +1,13 @@
 const webdriver = require('selenium-webdriver');
 const firefox = require('selenium-webdriver/firefox');
+const {Origin} = require('selenium-webdriver/lib/input');
 
 const {assert} = require('chai');
+
+const uuid = require('uuid');
+
+const firefoxExtensionId = '{68f0c654-5a3d-423b-b846-2b3ab68d05dd}';
+const extensionUuid = uuid.v4();
 
 function sleep(milliseconds) {
     return new Promise((resolve, reject) => {
@@ -21,6 +27,16 @@ function arraysEqual(arr1, arr2) {
     return true;
 }
 
+function eltOrNull(promise) {
+    return promise.catch((error) => {
+        if (error instanceof webdriver.error.NoSuchElementError) {
+            return null;
+        } else {
+            throw error;
+        }
+    });
+}
+
 class Link {
     constructor(id, element) {
         this.id = id;
@@ -38,16 +54,8 @@ class Link {
 
     async getTagline() {
         if (this.tagline === undefined) {
-            try {
-                this.tagline = await this.element.findElement(
-                    {css: '.tagline .rededup-tagline'});
-            } catch (error) {
-                if (error instanceof webdriver.error.NoSuchElementError) {
-                    this.tagline = null;
-                } else {
-                    throw error;
-                }
-            }
+            this.tagline = await eltOrNull(this.element.findElement(
+                {css: '.tagline .rededup-tagline'}));
         }
         return this.tagline;
     }
@@ -55,6 +63,11 @@ class Link {
     async getToggle() {
         const tagline = await this.getTagline();
         return await tagline.findElement({css: '.rededup-toggle'});
+    }
+
+    async getHash() {
+        return await eltOrNull(this.element.findElement(
+                {css: '.tagline .rededup-hash'}));
     }
 }
 
@@ -142,16 +155,18 @@ async function verifyShowHide(link, ...duplicateLinks) {
                  `Expect updated toggle text for link ${link.id}`);
 }
 
-async function deduplicateTest(driver, ids, expected) {
+async function deduplicateTest(driver, ids, expected, showHide = false) {
     const links = await loadByIds(driver, ids);
     await verifyDuplicates(links, expected);
-    let idx = 0;
-    for (const item of ids) {
-        if (Array.isArray(item)) {
-            await verifyShowHide(...links.slice(idx, idx+item.length));
-            idx += item.length;
-        } else {
-            idx++;
+    if (showHide) {
+        let idx = 0;
+        for (const item of ids) {
+            if (Array.isArray(item)) {
+                await verifyShowHide(...links.slice(idx, idx+item.length));
+                idx += item.length;
+            } else {
+                idx++;
+            }
         }
     }
     return links;
@@ -159,6 +174,10 @@ async function deduplicateTest(driver, ids, expected) {
 
 async function getVisibility(links) {
     return await Promise.all(links.map((link) => link.isVisible()));
+}
+
+async function openSettingsPage(driver) {
+    await driver.get(`moz-extension://${extensionUuid}/options/index.html`);
 }
 
 suite('Selenium', function() {
@@ -171,6 +190,9 @@ suite('Selenium', function() {
         if (process.env.BROWSER_GUI !== 'true') {
             options.headless();
         }
+        options.setPreference('extensions.webextensions.uuids',
+                              JSON.stringify({[firefoxExtensionId]:
+                                              extensionUuid}))
 
         driver = await new webdriver.Builder()
             .forBrowser('firefox')
@@ -186,19 +208,22 @@ suite('Selenium', function() {
         test('deduplicate by thumbnail', async function() {
             await deduplicateTest(driver,
                 ['t3_jrjed7', 't3_jrqhj4', 't3_jo1qwh', 't3_jri2y8'],
-                [['t3_jrjed7', 't3_jrqhj4', 't3_jri2y8'], 't3_jo1qwh']);
+                [['t3_jrjed7', 't3_jrqhj4', 't3_jri2y8'], 't3_jo1qwh'],
+                true);
         });
 
         test('deduplicate by url', async function() {
             await deduplicateTest(driver,
                 ['t3_jyu5b2', 't3_jysgvx', 't3_jywerx'],
-                [['t3_jyu5b2', 't3_jywerx'], 't3_jysgvx']);
+                [['t3_jyu5b2', 't3_jywerx'], 't3_jysgvx'],
+                true);
         });
 
         test('deduplicate crosspost', async function() {
             await deduplicateTest(driver,
                 ['t3_jyu5b2', 't3_jyvuiz'],
-                [['t3_jyu5b2', 't3_jyvuiz']]);
+                [['t3_jyu5b2', 't3_jyvuiz']],
+                true);
         });
 
         test('deduplicate multiple', async function() {
@@ -208,8 +233,10 @@ suite('Selenium', function() {
                 ['t3_jysgvx',
                  ['t3_jrjed7', 't3_jrqhj4', 't3_jri2y8'],
                  ['t3_jyu5b2', 't3_jywerx', 't3_jyvuiz'],
-                 't3_jo1qwh']);
+                 't3_jo1qwh'],
+                true);
 
+            // Verify that different toggle links function independently
             const toggle1 = await links[1].getToggle();
             const toggle2 = await links[4].getToggle();
 
@@ -248,6 +275,131 @@ suite('Selenium', function() {
             await click1();
             await click2();
             await click2();
+        });
+    });
+
+    suite('settings', function() {
+
+        test('deduplicate by', async function() {
+            // Posts should be collated
+            await deduplicateTest(driver,
+                ['t3_jrjed7', 't3_jrqhj4',],
+                [['t3_jrjed7', 't3_jrqhj4']]);
+
+            // Select "Deduplicate posts by URL only"
+            await openSettingsPage(driver);
+            const elt = await driver.findElement({id: 'ddByUrlOnly'});
+            await elt.click();
+
+            // Now posts should NOT be collated
+            await deduplicateTest(driver,
+                ['t3_jrjed7', 't3_jrqhj4',],
+                ['t3_jrjed7', 't3_jrqhj4']);
+        });
+
+        test('domain overrides', async function() {
+            // Posts should be collated
+            await deduplicateTest(driver,
+                ['t3_narqjh', 't3_na423x', 't3_nas73l'],
+                [['t3_narqjh', 't3_na423x', 't3_nas73l']]);
+
+            // Disable thumbnail processing for a domain
+            await openSettingsPage(driver);
+            let elt = await driver.findElement({id: 'domainSettingsButton'});
+            await elt.click();
+            elt = await driver.findElement({id: 'domainInputText'});
+            await elt.sendKeys('independent.co.uk');
+            elt = await driver.findElement({id: 'domainInputButton'});
+            await elt.click();
+            elt = await driver.findElement({id: 'domainSettingsSave'});
+            await elt.click();
+
+            // Now posts should be collated by URL but not thumbnail
+            await deduplicateTest(driver,
+                ['t3_narqjh', 't3_na423x', 't3_nas73l'],
+                [['t3_narqjh', 't3_nas73l'], 't3_na423x']);
+        });
+
+        for (const [name, eltId] of [['dct', 'dctHash'],
+                                     ['difference', 'diffHash'],
+                                     ['wavelet', 'waveletHash']]) {
+            test(`${name} hash function`, async function() {
+                // Select hash function
+                await openSettingsPage(driver);
+                const elt = await driver.findElement({id: eltId});
+                await elt.click();
+
+                // Verify posts are collated
+                await deduplicateTest(driver,
+                    ['t3_k7rhax', 't3_k7rdve'],
+                    [['t3_k7rhax', 't3_k7rdve']]);
+            });
+        }
+
+        test('hamming distance', async function() {
+            // Posts should all be collated
+            await deduplicateTest(driver,
+                ['t3_it6czg', 't3_it6el1', 't3_it8ric'],
+                [['t3_it6czg', 't3_it6el1', 't3_it8ric']]);
+
+            await openSettingsPage(driver);
+            const elt = await driver.findElement({id: 'maxHammingDistance'});
+            const {width} = await elt.getRect();
+            const actions = driver.actions();
+            await actions.move({origin: elt, x:0, y:0})
+                         .press()
+                         .move({origin: Origin.POINTER, x:-width/2, y:0})
+                         .release()
+                         .perform();
+            assert.equal(await elt.getAttribute('value'), '0');
+
+            // One post should now be separate from the other two
+            await deduplicateTest(driver,
+                ['t3_it6czg', 't3_it6el1', 't3_it8ric'],
+                [['t3_it6czg', 't3_it6el1'], 't3_it8ric']);
+        });
+
+        test('partition by domain', async function() {
+            // Posts with different domains should NOT be collated by default
+            await deduplicateTest(driver,
+                ['t3_jyymza', 't3_jyyy2a'],
+                ['t3_jyymza', 't3_jyyy2a']);
+
+            await openSettingsPage(driver);
+            const elt = await driver.findElement({id: 'partitionByDomain'});
+            await elt.click();
+
+            // Posts should now be collated
+            await deduplicateTest(driver,
+                ['t3_jyymza', 't3_jyyy2a'],
+                [['t3_jyymza', 't3_jyyy2a']]);
+        });
+
+        test('show hash values', async function() {
+            await openSettingsPage(driver);
+            const elt = await driver.findElement({id: 'showHashValues'});
+            await elt.click();
+
+            const links = await loadByIds(driver, ['t3_jyu5b2', 't3_ncjf1w']);
+            await verifyDuplicates(links, ['t3_jyu5b2', 't3_ncjf1w']);
+
+            assert.isNull(await links[0].getHash(),
+                          "Expect no thumbnail hash element for post without "
+                          + "thumbnail");
+
+            const hashElt = await links[1].getHash();
+            assert.isNotNull(hashElt,
+                             "Expect thumbnail hash for post with thumbnail");
+            // Use textContent to preserve whitespace
+            assert.match(await hashElt.getAttribute('textContent'),
+                         /^ \[[0-9a-f]{16}\]$/,
+                        "Expect hash text to be 16 hex digits in brackets");
+        });
+
+        teardown(async function() {
+            await openSettingsPage(driver);
+            const resetButton = driver.findElement({id: 'reset'});
+            await resetButton.click();
         });
     });